refactor(session): split transcript helpers into domain modules
Keep the @/lib/session barrel so callers and tests stay stable while types, tools, stream, rewind, and errors live in src/lib/session/.
